explicitly parametrize detail::deref calls to avoid unintentional deref
This is the main fix for unique_ptr<T> as non-unique-usertype: The code
correctly finds its operator< overload, but trying to instantiate it
accidentally dereferences it to T, which does not possess this overload.

None of the other calls to detail::deref looked like they would lead to
this issue, although it might not hurt to double-check.