Inside sol::proxy::call, check if we're building with clang
If building with clang on Windows, it tries to imitate MSVC by specifying reasonable _MSC_VER defines.
This makes sol2 compile MSVC version of sol::proxy::call (without the "template" preffix) that is nevertheless not accepted by clang.

This commit fixes the problem by also checking if "__clang__" is defined.